chore: add request rsponse schemas and segregate common auth and body
