+++ packages/hoppscotch-data/src/rest-request-response/v/0.ts
@@ -1,7 +1,24 @@
 import { defineVersion, entityReference } from "verzod"
 import { z } from "zod"
-import { HoppRESTResponseHeaders, ValidCodes } from "../../rest/v/8"
 import { HoppRESTResponseOriginalRequest } from "../original-request"
+import { StatusCodes } from "../../utils/statusCodes"
+
+export const ValidCodes = z.union(
+  Object.keys(StatusCodes).map((code) => z.literal(parseInt(code))) as [
+    z.ZodLiteral<number>,
+    z.ZodLiteral<number>,
+    ...z.ZodLiteral<number>[]
+  ]
+)
+
+export const HoppRESTResponseHeaders = z.array(
+  z.object({
+    key: z.string(),
+    value: z.string(),
+  })
+)
+
+export type HoppRESTResponseHeader = z.infer<typeof HoppRESTResponseHeaders>
 
 export const V0_SCHEMA = z.object({
   name: z.string(),
