+++ packages/hoppscotch-data/src/rest/v/11/index.ts
@@ -1,25 +1,24 @@
-import { defineVersion } from "verzod"
 import { z } from "zod"
 
-import { V2_SCHEMA } from "./2"
-import { HoppRESTAuth } from "../../../rest/v/11/auth"
-import { HoppRESTReqBody } from "../../../rest/v/10/body"
+import { V10_SCHEMA } from "../10"
+import { defineVersion } from "verzod"
 
-export const V3_SCHEMA = V2_SCHEMA.extend({
-  v: z.literal("3"),
+import { HoppRESTAuth } from "./auth"
+
+export const V11_SCHEMA = V10_SCHEMA.extend({
+  v: z.literal("11"),
   auth: HoppRESTAuth,
-  body: HoppRESTReqBody,
 })
 
 export default defineVersion({
+  schema: V11_SCHEMA,
   initial: false,
-  schema: V3_SCHEMA,
-  up(old: z.infer<typeof V2_SCHEMA>) {
+  up(old: z.infer<typeof V10_SCHEMA>) {
     const auth = old.auth
 
     return {
       ...old,
-      v: "3" as const,
+      v: "11" as const,
       auth:
         auth.authType === "oauth-2"
           ? {
