+++ packages/hoppscotch-data/src/rest/v/14.ts
@@ -1,12 +1,10 @@
 import { z } from "zod"
 import { defineVersion } from "verzod"
 import { V13_SCHEMA } from "./13"
-import { HoppRESTRequestResponses } from "../../rest-request-response"
 
 // Update the HoppRESTRequestResponses while migrating HoppRESTRequest
 export const V14_SCHEMA = V13_SCHEMA.extend({
   v: z.literal("14"),
-  responses: HoppRESTRequestResponses,
 })
 
 export default defineVersion({
