+++ packages/hoppscotch-data/src/rest/v/9/body.ts
@@ -1,6 +1,38 @@
 import { z } from "zod"
-import { FormDataKeyValue, V9_SCHEMA } from "./9"
-import { defineVersion } from "verzod"
+
+export const FormDataKeyValue = z
+  .object({
+    key: z.string(),
+    active: z.boolean(),
+    contentType: z.string().optional().catch(undefined),
+  })
+  .and(
+    z.union([
+      z.object({
+        isFile: z.literal(true),
+        value: z.array(z.instanceof(Blob).nullable()).catch([]),
+      }),
+      z.object({
+        isFile: z.literal(false),
+        value: z.string(),
+      }),
+    ])
+  )
+  .transform((data) => {
+    // Sample use case about restoring the `value` field in an empty state during page reload
+    // for files chosen in the previous attempt
+    if (data.isFile && Array.isArray(data.value) && data.value.length === 0) {
+      return {
+        ...data,
+        isFile: false,
+        value: "",
+      }
+    }
+
+    return data
+  })
+
+export type FormDataKeyValue = z.infer<typeof FormDataKeyValue>
 
 export const HoppRESTReqBody = z.union([
   z.object({
@@ -11,16 +43,10 @@ export const HoppRESTReqBody = z.union([
     contentType: z.literal("multipart/form-data"),
     body: z.array(FormDataKeyValue).catch([]),
     showIndividualContentType: z.boolean().optional().catch(false),
-    isBulkEditing: z.boolean().optional().catch(false),
   }),
   z.object({
     contentType: z.literal("application/octet-stream"),
     body: z.instanceof(File).nullable().catch(null),
-  }),
-  z.object({
-    contentType: z.literal("application/x-www-form-urlencoded"),
-    body: z.string().catch(""),
-    isBulkEditing: z.boolean().optional().catch(false),
   }),
   z.object({
     contentType: z.union([
@@ -30,6 +56,7 @@ export const HoppRESTReqBody = z.union([
       z.literal("application/vnd.api+json"),
       z.literal("application/xml"),
       z.literal("text/xml"),
+      z.literal("application/x-www-form-urlencoded"),
       z.literal("binary"),
       z.literal("text/html"),
       z.literal("text/plain"),
@@ -39,19 +66,3 @@ export const HoppRESTReqBody = z.union([
 ])
 
 export type HoppRESTReqBody = z.infer<typeof HoppRESTReqBody>
-
-export const V10_SCHEMA = V9_SCHEMA.extend({
-  v: z.literal("10"),
-  body: HoppRESTReqBody,
-})
-
-export default defineVersion({
-  schema: V10_SCHEMA,
-  initial: false,
-  up(old: z.infer<typeof V9_SCHEMA>) {
-    return {
-      ...old,
-      v: "10" as const,
-    }
-  },
-})
